Changed Game Screen Text to all have shadows + Adjusted Distance between multiple spikes
The "= 10 Points" was using TextMeshPro - Text, and the others were TextmeshPro - Text (UI), so I changed the rest of the text to use TextMeshPro - Text to allow the rest of the text on the Game Screen to have shadows and match, to make it look like it is hovering above the screen and make it look like it is a part of the game.

I also changed the spacing of the spikes so they weren't so far apart. If there are two or more spikes, they will now be touching each other so it looks like one set of spikes. This will also make the game a bit easier as it progresses as there will be more room between the spikes (they aren't as spread out).  I think this is good though as I found that the spikes being spread out far tended to make it pretty difficult relatively quickly.